Fix compilation error for kernel versions greater than V5.13
In V5.13 the support for Cyclades devices was removed from the kernel tree.

Commit: f76edd8f7ce06cdff2fe5b6b39a49644c684a161

The compilation error was caused because the macro MEN_UART_MAJOR was not
defined. To keep the compatilibity with systems that use the driver with
CYCLADES_MAJOR number, MEN_UART_MAJOR is defined with the same value.